Add Serialization in case1.
